Authentication JWT

+++ src/controllers/ProjectController.ts
@@ -5,6 +5,10 @@ export class ProjectController {
 
     static createProject = async (req: Request, res: Response) => {
         const project = new Project(req.body)
+
+        //Asigna un manager
+        project.manager = req.user.id
+
         try {
             await project.save()
             res.send('Proyecto Creado Correctamente')
@@ -15,7 +19,11 @@ export class ProjectController {
 
     static getAllProjects = async (req: Request, res: Response) => {
         try {
-            const projects = await Project.find({})
+            const projects = await Project.find({
+                $or: [
+                    { manager: {$in: req.user.id}}
+                ]
+            })
 
             if (projects.length === 0) {
                 const error = new Error('No hay proyectos')
@@ -39,6 +47,11 @@ export class ProjectController {
                 return res.status(404).json({ error: error.message })
             }
 
+            if (project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Accion no válida')
+                return res.status(404).json({error: error.message})
+            }
+
             res.json(project)
         } catch (error) {
             console.log(error)
@@ -54,6 +67,11 @@ export class ProjectController {
             if (!project) {
                 const error = new Error('Proyecto no Encontrado')
                 return res.status(404).json({ error: error.message })
+            }
+
+            if (project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Solo el manager puede actualizar un proyecto')
+                return res.status(404).json({error: error.message})
             }
 
             project.clientName = req.body.clientName
@@ -77,6 +95,11 @@ export class ProjectController {
             if (!project) {
                 const error = new Error('Proyecto no Encontrado')
                 return res.status(404).json({ error: error.message })
+            }
+
+            if (project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Solo el manager puede eliminar un proyecto')
+                return res.status(404).json({error: error.message})
             }
 
             await project?.deleteOne()
